third_party/WebKit/LayoutTests/http/tests/security/vibration/vibrate-in-same-origin-iframe-with-user-gesture-allowed-expected.txt

CONSOLE ERROR: line 19: Blocked call to navigator.vibrate because user hasn't tapped on the frame or any embedded frame yet: https://www.chromestatus.com/feature/5644273861001216.
This tests that a same-origin iframe with user gesture can vibrate.

127.0.0.1


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
This is a testharness.js-based test.
PASS Testing vibrate 3 times in an iframe
Harness: the test ran to completion.
